+++ ga.py
@@ -1,5 +1,6 @@
 import numpy as np
-from data_structure import SatInfo, fitness_func_with_param, TEST_SET, hash_function
+from data_structure import SatInfo, fitness_func_with_param
+from utils import TEST_SET, hash_function
 from sko.GA import GA
 import pandas as pd
 import matplotlib.pyplot as plt
